expenses dialog to expenses handler
